aria/requirements.txt
azure-cosmos==4.15.0
aiohttp==3.9.5
azure-servicebus==7.14.3
azure-functions==1.23.0
httpx==0.28.1
duckduckgo-search==8.1.1
nltk==3.9.3
scikit-learn==1.7.2
xgboost==3.2.0
networkx==3.6.1
numpy==2.4.3
pandas==3.0.1
python-dotenv==1.2.2
